Truncate the end of long token titles
By changing the autoresize rules on the token label, we can force the
label to appropriately size itself to the token title. Before, a
token’s label was sized to its content, meaning that if it was actually
longer than a line in the token field, it freely overflowed and was
clipped by the outer VENToken’s view. I also exposed the line break
mode for VENTokens as a whole, allowing a client application to change
the line break mode to one appropriate to their data (breaking in the
middle or beginning of the line, for example.