Simplify color scheme to use tint color
Rather than having an internal color scheme which can’t be easily
inherited, move to using the tintColor of the VENTokenField to control
the color of tokens. This allows just as much customizability, while
allowing common color schemes in an application to carry over without
custom color assignments. For example, changing the global tint color
will now be inherited by the token field if changed in either the
storyboard or by setting the UIWindow’s tint color. As a side effect,
tint adjustment modes (fading colors when an alert view is displayed)
is supported.