Fix layout issue causing the token field to be too tall
Because reloadData was being called before layoutScrollView, the scroll
view had a 0 width when attempting to calculate an initial height. This
meant that the first time the height was recalculated after being
displayed it would resize to the initially miscalculated 2 lines of
tokens height. Since there was no need for this, the input field stayed
on the first line, and there was a much larger section for the token
field than necessary. By skipping the setter call and just addressing
the label text directly, we avoid kicking off a data reload and its
associated layout run. Later changes (after rendering) will still cause
a data reload, preventing the client from having to manually reload any
time they change the prompt text.