Fix a layout bug related to the autocomplete table
By adding the table to the active window rather than to the
tokenField’s super view, I’m avoiding a bug where the frame  of the
token field is reset when it should be growing. This should also help
make sure the table view is on top of any other views that happen to
exist in the super view.