WIP: Allow selecting from autocomplete options
I made the decision that autocomplete is just a typing accelerator,
meaning that there is no dedicated callback for selecting an
autocompleted value, instead the actual value selected is sent back
just as if it was typed manually. I may change this before everything
is complete, as this pushes more (potential) verification on the client
without a lot saved for us.