Redesign autocomplete API
Instead of having an API that involves handing over a full array (that
may be thousands of items long) I switched over to one more similar to
a table view. I also broke out all of the autocomplete stuff to a
separate protocol instead of expanding the current data source.