Add seperate call back for autocomplete
Rather than making the client distinguish whether the entered text came
from a suggestion or not, we now have a delegate method for having
selected a suggestion. This is fired in addition to
tokenField:didEnterText: in order to allow users to opt in or to simply
enable autocomplete and use the same delegate methods as before.