Rename test to fix order
XCTest runners test in what appears to be alphabetical order, or
something close to it. Because of the naming, this test runs after
resignFirstResponder, which leaves the system in a state where that
test’s preconditions aren’t met - the keyboard is not up and ready to
input in the text field. By changing from assuming the first responder
to entering into the known good field, we stop worrying about order. I
also applied this fix to the testBasicFlow test, which had to be broken
into two steps due to the token field clearing properties.