Suppress NS_DESIGNATED_INITIALIZER warning
`-[UIView initWithFrame:]` is annotated with the
`NS_DESIGNATED_INITIALIZER` macro so it is expected that subclasses
will invoke the superclass’s implementation. Because we load our user
interface from a nib but still want to allow users to initialize an
instance with `-initWithFrame:`, we do not call super and instead
assign `self` to the result of unarchiving the aforementioned nib. This
causes a warning to be emitted about not calling `super`.

Since we know what we’re doing, we can work around this warning by just
calling `super` before we unarchive from the nib to appease the
compiler.