Delegate for return key when not completing token
When the user has the VENTokenField focused, and only has completed tokens, or no text at, tapping the return key now calls a new delegate method `tokenFieldDidReturn:`.

This is useful for the user to confirm entry of tokens, allowing the app to move focus to another text field or view.

`tokenField:didEnterText:` functionality has no change, and the code is written such that these two delegate methods are never called simultaneously.