Typing when token is selected now adds input text
Before this change, when you type while a token is selected, it adds text to the invisibleTextField, with no user feedback.  If you type multiple characters, you have to delete those characters before the token is deleted.  Now if you type while a token is selected, it deselects the token and adds the text to the visible inputTextField.